# openapi/momentom_api_openapi_1.0.1.yaml
openapi: 3.1.0
info:
  title: Momentom API (MVP)
  version: 1.0.1
  description: >
    Thin, explainable loop: ingest → readiness → adapt → explain → write-back.
    Contracts match UI Integration Spec (fields/states), Metrics & Fuelling, Field Spec, and AI/ML Display.
servers:
  - url: https://api.momentom.app/v1
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...
  schemas:
...
    ReadinessDriver:
      type: object
      properties:
        key: { type: string, enum: [hrv, rhr, sleep, soreness, mood, prior_strain, context] }
        z: { type: number, description: "z-score vs athlete baseline" }
        weight: { type: number }
        contribution: { type: number, description: "points +/- to composite" }
...
    DiffChange:
      type: object
      properties:
        op: { type: string, enum: [add, remove, replace] }  # nit: drop 'move'; express moves as replace on path with from/to
        path: { type: string, description: "JSON Pointer to field being changed" }
        from: {}
        to: {}
    Adaptation:
      type: object
      properties:
        adaptation_id: { type: string, format: uuid }
        plan_id: { type: string, format: uuid }
        scope: { type: string, enum: [today, next_72h, week] }
        impact_window:
          type: object
          properties:
            start: { type: string, format: date-time }
            end: { type: string, format: date-time }
        reason_code:
          type: string
          enum: [low_readiness, missed_session, monotony_high, ramp_high, illness]
        triggers:
          type: array
          items:
            type: string
            enum: [low_readiness, missed_session, monotony_high, ramp_high, illness]
        changes:
          type: array
          items:
              $ref: '#/components/schemas/DiffChange'
        decision: { type: string, enum: [proposed, accepted, modified, rejected] }
        plan_version_before: { type: integer, nullable: true }
        plan_version_after: { type: integer, nullable: true }
        rationale:
          type: object
          properties:
            text: { type: string }
            driver_attribution:
              type: array
              items:
              $ref: '#/components/schemas/ReadinessDriver'
            data_snapshot: { type: object, additionalProperties: true }
...
paths:
...
  /adaptations/preview:
    post:
      tags: [Adaptations]
      summary: Compute adaptation preview (no write)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                date: { type: string, format: date }
                scope: { type: string, enum: [today, next_72h, week] }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Adaptation' } } } }
...